# GetBBox performance improvements