--- a/test/all_tests.html
+++ b/test/all_tests.html
@@ -12,6 +12,7 @@
   <iframe src='contextmenu_test.html' width='100%' height='70' scrolling='no'></iframe>
   <iframe src='math_test.html' width='100%' height='70' scrolling='no'></iframe>
   <iframe src='svgutils_test.html' width='100%' height='70' scrolling='no'></iframe>
+  <iframe src='svgutils_bbox_test.html' width='100%' height='70' scrolling='no'></iframe>
   <iframe src='history_test.html' width='100%' height='70' scrolling='no'></iframe>
   <iframe src='select_test.html' width='100%' height='70' scrolling='no'></iframe>
   <iframe src='draw_test.html' width='100%' height='70' scrolling='no'></iframe>