… test/qunit/qunit-assert-close.js
… test/svgutils_bbox_test.html
… test/svgutils_performance_test.html
… test/svgutils_test.html