We just updated our project to the latest svg-edit (last update was three years ago). We've seen some major performance problems when performing selections in Chrome, Firefox, and Safari. With a large SVG document (270 top-level elements), select-all takes over 1 minute. This is especially evident when the top-level selections have matrix transforms.  When there are a few elements, we don't see the problem.

See [issue 26](https://github.com/SVG-Edit/svgedit/issues/26). We see the problems in the demo version pointed to from the svgedit page.

Chrome's profiler points to the native getBBox() taking all the time.

**The problems are due to:**

1) Both the thumbnail preview window and the prevalence of elements with matrix transforms (in our sample document). When getStrokedBBox/getCheckedBBox() sees a matrix transform, it calculates the BBox by enclosing a clone of the element and adding it to the DOM, calling getBBox(), then removing it. Without the transforms, there's no DOM change, so it's faster. If you delete the thumbnail and keep the transforms, it's faster, but still slow.

2) The various getBBox utility functions call the native getBBox() function multiple times for each element. It's at least 4 times for each top-level element when performing a rubber band selection.

3) svgcanvas: addToSelection() calls svgedit.utilities.getBBox(), throws the bbox away, then calls selectorManager.requestSelector() which calls svgedit.utilities.getBBox() again.

I've worked this week to refactor the related functions to calculate the BBox without updating the DOM and only call the native getBBox() once. I hope it's satisfactory. I've added a lot of tests and tested in Chrome, Firefox, and Safari on the latest OS X.

**Test Notes:**
* Please see the new performance test I added, svgutils_performance_test.html. I didn't add it to all_tests because I'm not sure what performance is expected for all platforms and devices.
* There are two new tests in svgedit.utilities_bbox that fail in Safari only. The last case in each deals with a rotated ellipse. I believe the BBox algorithm is incorrectly including the bezier control points to calculate the bounding box. i believe the Safari result is the correct one, but I'm not sure what is expected.
  * Test getBBoxWithTransform and a rotation transform.
  * Test getBBoxWithTransform with rotation and matrix transforms 

We're using svgedit for a schematic design tool (electric utility one-line diagrams).