@flintobrien 2016-04-25
Here are the performance improvement times with various browsers. The old code takes 4000 milliseconds to select-all and the new code is 34 milliseconds.

Test case test/svgutils_performance_test.html run on Macbook 2011 8MB RAM OS X El Capitan 10.11.4.
"Pass2" (below) is what you normally see when using svgedit to start a rubber band selection. Note: starting a rubber band selection calls getBBox on all objects.

#### Chrome
* Before Optimization
  * Pass1 svgCanvas.getStrokedBBox total ms 4,218, ave ms 41.0,   min/max 37 51
  * Pass2 svgCanvas.getStrokedBBox total ms 4,458, ave ms 43.3,   min/max 32 63
* Optimized Code
  * Pass1 svgCanvas.getStrokedBBox total ms 1,112, ave ms 10.8,   min/max 9 20
  * Pass2 svgCanvas.getStrokedBBox total ms    34, ave ms  0.3,   min/max 0 20

#### Firefox
* Before Optimization
  * Pass1 svgCanvas.getStrokedBBox total ms 3,794, ave ms 36.8,   min/max 33 48
  * Pass2 svgCanvas.getStrokedBBox total ms 4,049, ave ms 39.3,   min/max 28 53
* Optimized Code
  * Pass1 svgCanvas.getStrokedBBox total ms   104, ave ms 1.0,   min/max 0 23
  * Pass2 svgCanvas.getStrokedBBox total ms    71, ave ms 0.7,   min/max 0 23

#### Safari
* Before Optimization
  * Pass1 svgCanvas.getStrokedBBox total ms 4,840, ave ms 47.0,   min/max 45 62
  * Pass2 svgCanvas.getStrokedBBox total ms 4,849, ave ms 47.1,   min/max 34 62
* Optimized Code
  * Pass1 svgCanvas.getStrokedBBox total ms    42, ave ms 0.4,   min/max 0 23
  * Pass2 svgCanvas.getStrokedBBox total ms    17, ave ms 0.2,   min/max 0 23

If you want to run the performance test to see 'Before Optimization' yourself, make the following two edits to the new code.
1 svgedit.utilities.getStrokedBBox - change "if( elems.length === 1)" to "if( false && elems.length === 1)"
2 svgedit.utilities.getBBoxWithTransform - uncomment 'Old technique that was very slow'