@flintobrien 2016-04-25
I'm happy to do a phone call and walk through the changes. Whatever I can
do to help.
Regards,
Flint
919.260.8411


On Mon, Apr 25, 2016 at 12:09 PM, Jeff Schiller <notifications@github.com>
wrote:

> Wow - this sounds awesome - I'm hoping to take a look at this today or
> tonight.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/SVG-Edit/svgedit/pull/103#issuecomment-214421100>
>