@codedread 2016-04-26
Wow - thanks so much for this patch - I didn't look through the tests, just sort of skimmed through them.  Mostly I have concerns about passing svgcanvas functions into module functions - it's a bit of a hack, but frankly it's my fault for leaving the code in a half-migrated state where some of the code lives in separate modules and most of the code is still sitting in svgcanvas.js in a huge ball of spaghetti :)