@flintobrien 2016-04-26
On reflection....
If you're trying to line things up, stay on grid, or do precise resizing of a transformed element, it is paramount that the bbox be accurate (both the bounding rect and selection handles).