@codedread 2016-04-27
Ok, I see a lot of the other comments have been cleaned up - if you think this is ready, I will merge it... just let me know.