@flintobrien 2016-04-28
Thanks for the reviews. I think it's good to go with my last commit, [Optimized getBBoxWithTransform when rotation multiple of 90](https://github.com/SVG-Edit/svgedit/pull/103/commits/3230520d679a334cc1938dfbb5972783f63fd9b9).